optimal sampling (multiple m) now uses a convex optimization solver. very powerful!

+++ utility/parser.py
@@ -24,7 +24,7 @@ class ParserArgs(object):
         self.parser.add_argument("--powerfulCNN", action="store_true", help="Decide if use powerful CNN for EMNIST")
         self.parser.add_argument("--iid_type", nargs='*', default=['noniid', 'noniid', 'noniid', 'iid', 'iid'], help="List of type_iid")
         self.parser.add_argument("--task_type", nargs='*', default=['mnist', 'cifar10', 'fashion_mnist', 'emnist', 'cifar10'], help="List of task types")
-        self.parser.add_argument("--alpha", type=int, default=3, help="alpha value, alpha-fairness")
+        self.parser.add_argument("--alpha", type=float, default=3.0, help="alpha value, alpha-fairness")
         self.parser.add_argument("--notes", type=str, default='', help="notes to add on the folder name")
         self.parser.add_argument("--bayes_decay", type=float, default=1.0, help="Decay factor for bayesian method")
         self.parser.add_argument("--algo_type", nargs='*', default=['bayesian', 'proposed','random','round_robin'], help="List of algorithms")
